+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.5.7
 - Add **CodeLens** template.
+- Improve error location parsing.
 
 ## 0.5.6
 - Add **Delete** action to macro library nodes to unregister them.